requirements.txt
annotated-types==0.6.0
anyio==4.0.0
bcrypt==4.0.1
click==8.1.7
colorama==0.4.6
ecdsa==0.18.0
fastapi==0.104.0
greenlet==3.0.0
h11==0.14.0
idna==3.4
passlib==1.7.4
pip==23.2.1
pyasn1==0.5.0
pydantic-core==2.11.0
pydantic==2.4.2
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
setuptools==68.2.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.27.0
typing-extensions==4.8.0
uvicorn==0.23.2
wheel==0.41.2
asyncpg==0.28.0
